requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Computer Vision
opencv-python==4.8.0.76
dlib==19.24.1
numpy==1.25.2
scipy==1.11.2

# Data Processing
pandas==2.1.0
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyjwt==2.8.0
pytz==2023.3
Pillow==10.0.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.7.0

# Production
gunicorn==21.2.0 